+++ cookbooks/storm/recipes/nimbus.rb
@@ -1,0 +1,12 @@
+#
+# Cookbook Name:: storm
+# Recipe:: default
+#
+# Copyright 2012, Webtrends
+#
+# All rights reserved - Do Not Redistribute
+#
+
+include_recipe "storm"
+
+# do nimbus specific bits go here